src/requirements.txt
docopt==0.6.2
editdistance==0.5.3
joblib==0.16.0
jsonschema==3.2.0
lxml==4.5.2
matplotlib==3.3.1
nltk==3.5
numexpr==2.7.1
numpy
Pillow
psutil==5.7.2
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.2
tokenizers==0.8.1rc1
torch==1.6.0+cu102
torchvision==0.7.0+cu102
tqdm==4.48.2
transformers==3.0.2
typing-extensions==3.7.4.3
urllib3==1.24.3
ujson==3.1.0
